pyproject.toml
[tool.poetry]
name = "nal-backend"
version = "0.1.0"
description = "NAL Platform Backend Service"
authors = ["Abhishek Kumar Seth"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.3.2"
pydantic = "^2.11.9"
python-dotenv = "^1.1.1"
fastapi = "^0.117.1"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
uvicorn = "^0.37.0"
yarl = "^1.20.1"
pydantic-settings = "2.7.1"
loguru = "^0.7.3"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
toml = "^0.10.2"
httpx = "^0.28.1"
pyjwt = "^2.8.0"
phonenumbers = "^8.13.0"
bcrypt = "^4.0.0"
redis = "^5.0.0"
email-validator = "^2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<2.1.0"]
build-backend = "poetry.core.masonry.api"